composition_service/requirements.txt
Flask
python-dotenv
requests
grpcio
grpcio-tools
